backend/requirements.txt
fastapi>=0.103.0
uvicorn[standard]>=0.23.2
gunicorn>=21.2.0
asyncpg>=0.28.0
databases>=0.8.0
alembic>=1.12.0
google-generativeai>=0.4.0
anthropic>=0.18.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-anthropic>=0.1.0
pinecone-client>=3.0.0
redis>=5.0.1
aioredis>=2.0.1
pydantic>=2.4.0
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
sse-starlette>=1.8.2
openai>=1.0.0
